% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hellokit_package.R
\docType{package}
\name{hellokit-package}
\alias{hellokit-package}
\title{hellokit-package
A Good Title
A thorough description}
\description{
hellokit-package
A Good Title
A thorough description
}

